backend/requirements.txt
# StudyMate AI - Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# File Processing
pypdf>=4.0.0
python-docx>=1.1.0

# AI / ML - Embeddings
sentence-transformers>=2.3.0

# AI / ML - Vector Database
faiss-cpu>=1.7.4

# AI / ML - LLM (LangChain)
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0

# AI API SDKs
groq>=0.8.0
google-generativeai>=0.5.0
openai>=1.10.0

# Database (auth)
sqlalchemy>=2.0.0

# Password hashing
bcrypt>=4.0.0

# Email
resend>=0.7.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic[email]>=2.5.0
